backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.11.7
pydantic-settings==2.7.1
